feat(cs): Update files controllers and dto_s

+++ Proyecto-REMI-WebApi/Proyecto-REMI-WebApi/Models/DTO's/pedidoDetalleDto.cs
@@ -3,8 +3,8 @@
     public class pedidoDetalleDto
     {
         public int codigoProducto { get; set; }
-        public double cantidadProducto { get; set; }
-        public double valorProducto { get; set; }
-        public double totalPagoProducto { get; set; }
+        public int cantidadProducto { get; set; }
+        public decimal valorProducto { get; set; }
+        public decimal totalPagoProducto { get; set; }
     }
 }
